[OTA] Create recovery image
1. Extend /cache size to 48MB
2. Revise for recovery image, but it still cannot be loaded correctly

+++ meta-WISE-PaaS/recipes-ota/images/initramfs-debug-image.bbappend
@@ -1,9 +1,7 @@
 DEPNEDS += "${PREFERRED_PROVIDER_virtual/kernel}"
 
 BOOTIMG_PAGE_SIZE ?= "2048"
-CACHE_PARTITION_mx6 = "/dev/disk/by-label/cache"
-CACHE_PARTITION_dragonboard-410c = "/dev/disk/by-partlabel/cache"
-CACHE_PARTITION_arago = "/dev/disk/by-label/cache"
+CACHE_PARTITION = "/dev/disk/by-label/cache"
 
 modify_fstab() {
         echo "${CACHE_PARTITION}      /cache               ext4       nosuid,nodev,nomblk_io_submit 0 0" >> ${IMAGE_ROOTFS}/etc/fstab
@@ -11,46 +9,46 @@ modify_fstab() {
 
 ROOTFS_POSTPROCESS_COMMAND += " modify_fstab;"
 
+# *.cpio.gz is moved to deploy dir at sstate_task_postfunc() in Yocto 2.5
+# We have to generate recovery.img after that
+sstate_task_postfunc_append() {
+    from oe.utils import execute_pre_post_process
+    post_mk_recovery_cmds = d.getVar("RECOVERY_POSTPROCESS_COMMAND")
+    execute_pre_post_process(d, post_mk_recovery_cmds)
+}
+
 # [i.MX]
 DEPENDS_imx += "android-tools-native"
+BASE_ADDR_mx6 = "0x14000000"
+BASE_ADDR_mx8 = "0x82280000"
+SEC_OFFSET_mx6 = "0x00f00000"
+SEC_OFFSET_mx8 = "0x01f00000"
+RAMDISK_OFFSET_mx6 = "0x01000000"
+RAMDISK_OFFSET_mx8 = "0x02000000"
 
 mk_recovery_img_imx() {
+    if [ -e ${DEPLOY_DIR_IMAGE}/${PN}-${MACHINE}.cpio.gz ]; then
+        # Make sure only one dts files is selected
+        export ALL_DTS_FILE="${KERNEL_DEVICETREE}"
+        export FIRST_DTS_FILE=`echo ${ALL_DTS_FILE#*/} | cut -d ' ' -f 1`
 
-    mkbootimg --kernel ${DEPLOY_DIR_IMAGE}/${KERNEL_IMAGETYPE} \
-              --ramdisk ${DEPLOY_DIR_IMAGE}/${PN}-${MACHINE}.cpio.gz \
-              --output ${DEPLOY_DIR_IMAGE}/recovery-${MACHINE}.img \
-              --second ${DEPLOY_DIR_IMAGE}/${KERNEL_IMAGETYPE}-${KERNEL_DEVICETREE} \
-              --pagesize ${BOOTIMG_PAGE_SIZE} \
-              --base 0x14000000 \
-              --cmdline ""
+        gunzip -f -k ${DEPLOY_DIR_IMAGE}/${PN}-${MACHINE}.cpio.gz
 
-    ln -sf recovery-${MACHINE}.img ${DEPLOY_DIR_IMAGE}/recovery.img
+        mkbootimg --kernel ${DEPLOY_DIR_IMAGE}/${KERNEL_IMAGETYPE} \
+                  --ramdisk ${DEPLOY_DIR_IMAGE}/${PN}-${MACHINE}.cpio \
+                  --output ${DEPLOY_DIR_IMAGE}/recovery-${MACHINE}.img \
+                  --second ${DEPLOY_DIR_IMAGE}/${KERNEL_IMAGETYPE}-${FIRST_DTS_FILE} \
+                  --pagesize ${BOOTIMG_PAGE_SIZE} \
+                  --base ${BASE_ADDR} \
+                  --second_offset ${SEC_OFFSET} \
+                  --ramdisk_offset ${RAMDISK_OFFSET} \
+                  --cmdline ""
+
+        ln -sf recovery-${MACHINE}.img ${DEPLOY_DIR_IMAGE}/recovery.img
+    fi
 }
 
-IMAGE_POSTPROCESS_COMMAND_imx += " mk_recovery_img_imx ; "
-
-# [Qcom APQ8016]
-DEPENDS_dragonboard-410c += "skales-native"
-
-mk_recovery_img_qcom() {
-
-    export LD_LIBRARY_PATH="$LD_LIBRARY_PATH:${STAGING_LIBDIR_NATIVE}"
-    dtbTool -o ${DEPLOY_DIR_IMAGE}/dt-Image-${MACHINE}.img -s ${BOOTIMG_PAGE_SIZE} ${DEPLOY_DIR_IMAGE}
-    mkbootimg_dtarg="--dt ${DEPLOY_DIR_IMAGE}/dt-Image-${MACHINE}.img"
-    ln -sf dt-Image-${MACHINE}.img ${DEPLOY_DIR_IMAGE}/dt-Image.img
-
-    mkbootimg --kernel ${DEPLOY_DIR_IMAGE}/${KERNEL_IMAGETYPE} \
-              --ramdisk ${DEPLOY_DIR_IMAGE}/${PN}-${MACHINE}.cpio.gz \
-              --output ${DEPLOY_DIR_IMAGE}/recovery-${MACHINE}.img \
-              $mkbootimg_dtarg \
-              --pagesize ${BOOTIMG_PAGE_SIZE} \
-              --base 0 \
-              --cmdline ""
-
-    ln -sf recovery-${MACHINE}.img ${DEPLOY_DIR_IMAGE}/recovery.img
-}
-
-IMAGE_POSTPROCESS_COMMAND_dragonboard-410c += " mk_recovery_img_qcom ; "
+RECOVERY_POSTPROCESS_COMMAND_imx += " mk_recovery_img_imx;"
 
 # Utilities
 PACKAGE_INSTALL += " adv-ota android-tools base-files boottimes e2fsprogs-tune2fs udev-extraconf util-linux e2fsprogs-resize2fs e2fsprogs-e2fsck"
